+++ packages/core/src/clients/index.ts
@@ -13,7 +13,10 @@ export {
 	PresignUrlOptions,
 	SignRequestOptions,
 } from './middleware/signing/signer/signatureV4';
-export { EMPTY_HASH as EMPTY_SHA256_HASH } from './middleware/signing/signer/signatureV4/constants';
+export {
+	EMPTY_HASH as EMPTY_SHA256_HASH,
+	UNSIGNED_PAYLOAD,
+} from './middleware/signing/signer/signatureV4/constants';
 export { extendedEncodeURIComponent } from './middleware/signing/utils/extendedEncodeURIComponent';
 export {
 	signingMiddlewareFactory,