+++ packages/rtn-web-browser/android/src/main/kotlin/com/amazonaws/amplify/rtnwebbrowser/WebBrowserPackage.kt
@@ -18,5 +18,8 @@ class WebBrowserPackage : ReactPackage {
 
     override fun createNativeModules(
         reactContext: ReactApplicationContext
-    ): MutableList<NativeModule> = listOf(WebBrowserModule(reactContext)).toMutableList()
+    ): MutableList<NativeModule> = listOf(
+        WebBrowserModule(reactContext),
+        ChromeOSModule(reactContext)
+    ).toMutableList()
 }