+++ packages/rtn-push-notification/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/rtn-push-notification@1.2.35...@aws-amplify/rtn-push-notification@1.3.0) (2025-09-29)
+
+### Features
+
+- **rtn-passkeys:** add rtn passkeys ([#14489](https://github.com/aws-amplify/amplify-js/issues/14489)) ([ac06885](https://github.com/aws-amplify/amplify-js/commit/ac06885a8f2a13365fe1ff26bfd54891f9ad4a45)), closes [#14390](https://github.com/aws-amplify/amplify-js/issues/14390) [#14391](https://github.com/aws-amplify/amplify-js/issues/14391)
+
 ## [1.2.35](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/rtn-push-notification@1.2.34...@aws-amplify/rtn-push-notification@1.2.35) (2025-05-27)
 
 **Note:** Version bump only for package @aws-amplify/rtn-push-notification